Fix issue with branch not tracking to remote on first run
When a user runs `ninny stage_up` for the first time on a new staging
branch, it does not properly track the branch. The `branch.checkout`
method in ruby-git actually does this

```sh
git branch 'my_branch'
git checkout 'my_branch'
```

and that first line causes the new branch to be setup on whatever branch
the repo is currently on (not the remote branch). Later on, tracking is
set up against the remote branch, and it does a `git pull` which merges
the remote INTO the local.

Just running `git.checkout('my_branch')` seems to solve the problem, because
it does not run the `git branch 'my_branch'` before the checkout.